Move qth functions to qth source files. Move remaining tle-db-related functions to tle-file, ditto for transponder db functions.
