+++ flyby_ui.c
@@ -52,7 +52,7 @@
 #include <fcntl.h>
 #include <predict/predict.h>
 #include "flyby_ui.h"
-#include "flyby_config.h"
+#include "qth_config.h"
 
 #define EARTH_RADIUS_KM		6.378137E3		/* WGS 84 Earth radius km */
 #define HALF_DELAY_TIME	5
